Switch sequence-id to 32-bit from 64
I saw segfaults for nil deref when deploying on a linux_386 platform;
this is a deliberate decision in the stdlib sync/atomic package for
missing instructions, not documented that I can see, but reasonable.

We don't need 64-bits, it was just a reasonable default.  Switch to
uint32 for the seq-id.

+++ renew/cert.go
@@ -24,7 +24,7 @@ var (
 type CertRenewal struct {
 	*Renewer
 
-	ActionID    uint64
+	ActionID    uint32
 	actionIDStr string
 
 	certPath   string
